When font fails to load, grey out text instead of unmapping text widget
Before if you selected a pattern that couldn't load, such as a strange
encoding, the text widget would simply stop updating, giving the user
no feedback that they weren't seeing the selected font, and leaving
repaint turds from menus in the text area.   Now we are more clear that
something is wrong, and clean up when menus overlap the text.

Signed-off-by: Alan Coopersmith <alan.coopersmith@oracle.com>